refactor(shared): move location helpers out of lib
