+++ src/features/calendar/server/ical-section-exam-event.ts
@@ -13,8 +13,8 @@ import {
 } from "@/features/calendar/server/ical-labels";
 import type { Prisma } from "@/generated/prisma/client";
 import type { AppLocale } from "@/i18n/config";
-import { lookupBuildingImagePath } from "@/lib/location-utils";
 import { APP_TIME_ZONE } from "@/lib/time/parse-date-input";
+import { lookupBuildingImagePath } from "@/shared/lib/location/location-utils";
 
 export function createExamEvent(
   exam: Prisma.ExamGetPayload<{ include: { examRooms: true } }>,
