--- src/lib/location-utils.ts
@@ -1,10 +1,0 @@
-export type { ScheduleLocationInfo } from "@/lib/schedule-location-format";
-export { formatScheduleLocation } from "@/lib/schedule-location-format";
-export {
-  getBuildingImagePath,
-  getLocationGeo,
-  loadBuildingImgRules,
-  loadGeoData,
-  lookupBuildingImagePath,
-  lookupLocationGeo,
-} from "@/lib/static-location-data";
