+++ src/shared/lib/location/location-utils.ts
@@ -1,0 +1,10 @@
+export type { ScheduleLocationInfo } from "./schedule-location-format";
+export { formatScheduleLocation } from "./schedule-location-format";
+export {
+  getBuildingImagePath,
+  getLocationGeo,
+  loadBuildingImgRules,
+  loadGeoData,
+  lookupBuildingImagePath,
+  lookupLocationGeo,
+} from "./static-location-data";
