+++ src/shared/lib/location/schedule-location-format.ts
