+++ src/shared/lib/location/static-location-data.ts
@@ -2,14 +2,14 @@ export {
   getBuildingImagePath,
   loadBuildingImgRules,
   lookupBuildingImagePath,
-} from "@/lib/static-building-images";
+} from "./static-building-images";
 export {
   getLocationGeo,
   loadGeoData,
   lookupLocationGeo,
-} from "@/lib/static-geo-data";
+} from "./static-geo-data";
 export type {
   BuildingImgRule,
   GeoData,
   GeoLocation,
-} from "@/lib/static-location-types";
+} from "./static-location-types";
