+++ src/shared/lib/location/static-location-types.ts
